requirements.txt
weaviate-client<4.0
pydantic<2.0
langchain==0.0.27
tiktoken
requests 